ai-reviewer-tool/requirements.txt
# Core AI/ML dependencies
crewai==0.28.0
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-openai==0.0.5
langchain-community==0.0.10
langchain-huggingface==0.0.6
langchain-groq==0.0.3
transformers==4.36.0
torch==2.1.0

# Testing and development
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
flake8==6.1.0
black==23.12.0
mypy==1.8.0

# Authentication and security
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# HTTP and networking
requests==2.31.0
urllib3==2.1.0

# File and path handling
pathlib2==2.3.7
typing-extensions==4.9.0

# AI/ML specific
huggingface-hub==0.20.0
tokenizers==0.15.0

# Git operations
GitPython==3.1.40

# Environment and configuration
python-dotenv==1.0.0

# Data processing
pandas==2.1.4
numpy==1.24.3

# Logging and monitoring
structlog==23.2.0

# Additional utilities
tqdm==4.66.1
colorama==0.4.6 